+++ libsrc/filter.cc
@@ -37,9 +37,9 @@ void inflate(TriMesh *mesh, float amount)
 	mesh->need_normals();
 
 	dprintf("Creating offset surface... ");
-	int nv = mesh->vertices.size();
+	size_t nv = mesh->vertices.size();
 #pragma omp parallel for
-	for (int i = 0; i < nv; i++)
+	for (size_t i = 0; i < nv; i++)
 		mesh->vertices[i] += amount * mesh->normals[i];
 	dprintf("Done.\n");
 	mesh->bbox.valid = false;
@@ -50,16 +50,16 @@ void inflate(TriMesh *mesh, float amount)
 // Transform the mesh by the given matrix
 void apply_xform(TriMesh *mesh, const xform &xf)
 {
-	int nv = mesh->vertices.size();
+	size_t nv = mesh->vertices.size();
 
 #pragma omp parallel for
-	for (int i = 0; i < nv; i++)
+	for (size_t i = 0; i < nv; i++)
 		mesh->vertices[i] = xf * mesh->vertices[i];
 
 	if (!mesh->normals.empty()) {
 		xform nxf = norm_xf(xf);
 #pragma omp parallel for
-		for (int i = 0; i < nv; i++) {
+		for (size_t i = 0; i < nv; i++) {
 			mesh->normals[i] = nxf * mesh->normals[i];
 			normalize(mesh->normals[i]);
 		}
@@ -114,9 +114,9 @@ void scale(TriMesh *mesh, float s, const vec &d)
 // Clip mesh to the given bounding box 
 void clip(TriMesh *mesh, const box &b)
 {
-	int nv = mesh->vertices.size();
+	size_t nv = mesh->vertices.size();
 	vector<bool> toremove(nv, false);
-	for (int i = 0; i < nv; i++)
+	for (size_t i = 0; i < nv; i++)
 		if (mesh->vertices[i][0] < b.min[0] ||
 		    mesh->vertices[i][0] > b.max[0] ||
 		    mesh->vertices[i][1] < b.min[1] ||
@@ -147,8 +147,8 @@ point mesh_center_of_mass(TriMesh *mesh)
 	float totwt = 0;
 
 	mesh->need_faces();
-	int nf = mesh->faces.size();
-	for (int i = 0; i < nf; i++) {
+	size_t nf = mesh->faces.size();
+	for (size_t i = 0; i < nf; i++) {
 		const point &v0 = mesh->vertices[mesh->faces[i][0]];
 		const point &v1 = mesh->vertices[mesh->faces[i][1]];
 		const point &v2 = mesh->vertices[mesh->faces[i][2]];
@@ -169,8 +169,8 @@ void point_covariance(const vector<point> &pts, float C[3][3])
 		for (int k = 0; k < 3; k++)
 			C[j][k] = 0.0f;
 
-	int n = pts.size();
-	for (int i = 0; i < n; i++) {
+	size_t n = pts.size();
+	for (size_t i = 0; i < n; i++) {
 		for (int j = 0; j < 3; j++)
 			for (int k = j; k < 3; k++)
 				C[j][k] += pts[i][j] * pts[i][k];
@@ -202,8 +202,8 @@ void mesh_covariance(TriMesh *mesh, float C[3][3])
 
 	float totarea = 0.0f;
 	const vector<point> &p = mesh->vertices;
-	int n = mesh->faces.size();
-	for (int i = 0; i < n; i++) {
+	size_t n = mesh->faces.size();
+	for (size_t i = 0; i < n; i++) {
 		const TriMesh::Face &f = mesh->faces[i];
 		point c = (p[f[0]] + p[f[1]] + p[f[2]]) / 3.0f;
 		float area = len(trinorm(p[f[0]], p[f[1]], p[f[2]]));
@@ -265,8 +265,8 @@ void pca_rotate(TriMesh *mesh)
 	// Sorted in order from smallest to largest, so grab third column
 	vec first(C[0][2], C[1][2], C[2][2]);
 	int npos = 0;
-	int nv = mesh->vertices.size();
-	for (int i = 0; i < nv; i++)
+	size_t nv = mesh->vertices.size();
+	for (size_t i = 0; i < nv; i++)
 		if ((mesh->vertices[i] DOT first) > 0.0f)
 			npos++;
 	if (npos < nv/2)
@@ -308,8 +308,8 @@ void pca_snap(TriMesh *mesh)
 	// Sorted in order from smallest to largest, so grab third column
 	vec first(C[0][2], C[1][2], C[2][2]);
 	int npos = 0;
-	int nv = mesh->vertices.size();
-	for (int i = 0; i < nv; i++)
+	size_t nv = mesh->vertices.size();
+	for (size_t i = 0; i < nv; i++)
 		if ((mesh->vertices[i] DOT first) > 0.0f)
 			npos++;
 	if (npos < nv/2)
@@ -395,8 +395,8 @@ void orient(TriMesh *mesh)
 	dprintf("Auto-orienting mesh... ");
 	unsigned cc = 0;
 	vector<int> cc_farthest;
-	int nf = mesh->faces.size();
-	for (int i = 0; i < nf; i++) {
+	size_t nf = mesh->faces.size();
+	for (size_t i = 0; i < nf; i++) {
 		if (mesh->flags[i] != NONE)
 			continue;
 		mesh->flags[i] = cc;
@@ -471,9 +471,9 @@ void orient(TriMesh *mesh)
 // Remove boundary vertices (and faces that touch them)
 void erode(TriMesh *mesh)
 {
-	int nv = mesh->vertices.size();
+	size_t nv = mesh->vertices.size();
 	vector<bool> bdy(nv);
-	for (int i = 0; i < nv; i++)
+	for (size_t i = 0; i < nv; i++)
 		bdy[i] = mesh->is_bdy(i);
 	remove_vertices(mesh, bdy);
 }
@@ -484,14 +484,14 @@ void noisify(TriMesh *mesh, float amount)
 {
 	mesh->need_normals();
 	mesh->need_neighbors();
-	int nv = mesh->vertices.size();
+	size_t nv = mesh->vertices.size();
 	vector<vec> disp(nv);
 
-	for (int i = 0; i < nv; i++) {
+	for (size_t i = 0; i < nv; i++) {
 		point &v = mesh->vertices[i];
 		// Tangential
-		int nn = mesh->neighbors[i].size();
-		for (int j = 0; j < nn; j++) {
+		size_t nn = mesh->neighbors[i].size();
+		for (size_t j = 0; j < nn; j++) {
 			const point &n = mesh->vertices[mesh->neighbors[i][j]];
 			float scale = amount / (amount + len(n-v));
 			disp[i] += (float) tinyrnd() * scale * (n-v);
@@ -502,7 +502,7 @@ void noisify(TriMesh *mesh, float amount)
 		disp[i] += (2.0f * (float) tinyrnd() - 1.0f) *
 			   amount * mesh->normals[i];
 	}
-	for (int i = 0; i < nv; i++)
+	for (size_t i = 0; i < nv; i++)
 		mesh->vertices[i] += disp[i];
 }
